go2go: overhaul the Buffer and PubSub types to use generics
Use the go2go generics experiment to enforce type checking at compile
time instead of runtime.

Switch the Buffer type to manipulate a slice of comparable T's instead
of a slice of empty interfaces. Remove the global EmptyMarker in favor
of a per-Buffer instance empty marker value.

Add an any T type parameter to the PubSub type. Use an internal Buffer
of cells of T, which hold a T value or a marker channel. The marker
channels indicate an unsubscribing consumer or the closing of the
PubSub instance.

Drop the Publisher & Subscriber API, it is no longer needed now that the
publish & subscribe methods for both functions & channels support the
instance's type parameter.