Add arm64 Support
According to the  package [bug
notes](https://pkg.go.dev/sync/atomic#pkg-note-BUG), on arm64 platforms the
first word in an allocated struct is guaranteed to be 64 bit aligned. The
cursor and slice packages only operate on the first item in the Cursor struct,
so we should be good there.

I waffled a bunch on how to implement this, either making it the default, or
using a file with multiple build tags. I ended up just copying it in a new file
to make it clear that while the same code works, the two platforms have
different behaviour and any changes made should consider this.